Moved MiniRouter into separate directory
